requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
numpy==1.24.3
python-dotenv==0.19.0
Werkzeug==2.0.1
